Fix class extraction errors.
